Add environments for evaluation to PointMaze datasets (#3)
* add all dataset ids

* eval env

* single-goal, multi-reset maps

* increase checkpoint steps

* convert controller action to space

---------

+++ scripts/pointmaze/controller.py
@@ -51,5 +51,5 @@ class WaypointController:
 
         action = self.gains['p'] * (self.current_control_target_xy - obs['achieved_goal']) + self.gains['d'] * obs['observation'][2:]
         
-        return action
+        return np.float32(action)
